requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.13.3
trafilatura==1.12.2
youtube-transcript-api==0.6.3
yt-dlp>=2024.8.6
openai==1.65.2
python-dotenv==1.0.1
